requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-login
pymysql
werkzeug
